### 2023-06-19 #74 @Josuelmet
[Re] The Discriminative Kalman Filter for Bayesian Filtering with Nonlinear and Non-Gaussian Observation Models
**Original article:** The Discriminative Kalman Filter for Bayesian Filtering with Nonlinear and Non-Gaussian Observation Models

**Scientific domain:** Computational Neuroscience
