### 2023-12-21 #74 @Josuelmet
@benoit-girard Sure,
I think @ozancaglayan @mxochicale @gdetor would be good picks, since they are familiar with signal processing and computational neuroscience.
