### 2024-01-19 #74 @benoit-girard
@ozancaglayan Your status wrt. academia is not a problem (we don't have such requirements at ReScience C), as long as you have the expertise and know-how to perform the review.
It would be a pleasure to have you doing it, as I have been unsucessfully looking for reviewers so far.
