### 2024-02-11 #74 @ozancaglayan
# Review

First of all, I'd like to thank the authors as the paper is very clear and fun to read! The authors went for replicating the experiment 4.5 which they claim to be the most significant type of experiment for BCI studies. The replication study is able to confirm the trends shown by the original authors with one method underperforming and another outperforming the original results.

I was particularly curious about the LSTM/NN approach which often requires quite some tuning to get the best out of it. I'm pretty sure it could have surpassed the other methods given enough time and patience. Authors here already show that by changing a couple of settings, they were able to get better performance than the claimed ones.

Unfortunately, I'm having issues in running the experiments so let's just first focus on making this work and then I'll follow up with the replication status.

**Test environments:**
- Python 3.12 w/ Mambaforge + `pip install -r requirements.txt` (Ubuntu 22.04.3, x86_64) ❌
- Same as above but on Mac M1 fails ❌ (see below, this is more of a technical issue which should just be mentioned)

## Readability
- It would be very beneficial for readability if you use `black`, `isort` and friends to lint and format your `.py` scripts in the root repository.

- Please clear the cell outputs from the notebook files a bit. There are lots of warnings printed out there cluttering the actual output, you can use `with np.errstate(divide='ignore', invalid='ignore'):` wrappers around expected division issues:
```C:\Users\josue\Documents\Python Scripts\DKF\Python_Paper_45\nw_est.py:26: RuntimeWarning: invalid value encountered in true_divide
  f_est = xP / sP
C:\Users\josue\Documents\Python Scripts\DKF\Python_Paper_45\optimize_sx.py:42: RuntimeWarning: Mean of empty slice
  return np.nanmean((xtrain - preds) ** 2)
```
## Issues
- `sklearn` dependency should be renamed to `scikit-learn` in `requirements.txt`
```bash
  × python setup.py egg_info did not run successfully.
  │ exit code: 1
  ╰─> [15 lines of output]
      The 'sklearn' PyPI package is deprecated, use 'scikit-learn'
      rather than 'sklearn' for pip commands.
```

- There seems to be an issue on Macs: Jupyter kernel crashes. When I launch it from the CLI as a script, I get a warning & error regarding `Hint This means that multiple copies of the OpenMP runtime have been linked into the program.`. There are some [suggestions](https://stackoverflow.com/questions/53014306/error-15-initializing-libiomp5-dylib-but-found-libiomp5-dylib-already-initial) regarding this and one suggestion was to switch the imports for `torch` and `numpy` which seemed to make the error go away but now the notebook stuck at `Iteration 0`. Created a clean conda environment and installed the `nomkl` package first, and then the dependencies. Now I get a `segmentation fault`. It may be beneficial to add a note regarding this in README and recommend people to use `x86_64` architecture machines.

- I now switched to Linux. When running the actual Paper replication script, I get this:
```python
Iteration 0:
DKF-NN
/home/ozan.caglayan/Discriminative-Kalman-Filter-4.5-Python/optimize_sx.py:43: RuntimeWarning: Mean of empty slice
  return np.nanmean((xtrain - preds) ** 2)
DKF-GP
/home/ozan.caglayan/Discriminative-Kalman-Filter-4.5-Python/optimize_sx.py:43: RuntimeWarning: Mean of empty slice
  return np.nanmean((xtrain - preds) ** 2)
Traceback (most recent call last):
  File "/home/ozan.caglayan/Discriminative-Kalman-Filter-4.5-Python/Paper_Script_45.py", line 268, in <module>
    zpred.loc[iRun, "DKF_GP"] = DKF_filtering(
                                ^^^^^^^^^^^^^^
  File "/home/ozan.caglayan/Discriminative-Kalman-Filter-4.5-Python/DKF_filtering.py", line 30, in DKF_filtering
    if not np.all(eig(inv(bb) - inv(VZ))[0] > 0):
                  ^^^^^^^^^^^^^^^^^^^^^^
  File "/home/ozan.caglayan/mambaforge/envs/kalman/lib/python3.12/site-packages/numpy/linalg/linalg.py", line 1329, in eig
    _assert_finite(a)
  File "/home/ozan.caglayan/mambaforge/envs/kalman/lib/python3.12/site-packages/numpy/linalg/linalg.py", line 218, in _assert_finite
    raise LinAlgError("Array must not contain infs or NaNs")
numpy.linalg.LinAlgError: Array must not contain infs or NaNs
```

## Other
- You have the `.npy` files committed to the `flint-data-preprocessing` subfolder. Do we want this? As you already provide the pre-processing logic with freshly downloaded MAT files, it makes it a bit confusing for the user to understand what to expect here. Should they pre-process from scratch or use the provided files?

- In `flint-data-preprocessing/README.md`, you instruct the reader to move the `.npy` files to the root folder but further `ipynb` notebooks actually access them from the `flint-data-preprocessing` subfolder which makes sense. Could you remove that instruction from the README file maybe for clarity?
