### 2024-03-05 cc24938
Update requirements.txt
replaced sklearn with scikit-learn in requirements.txt
