+++ requirements.txt
@@ -3,5 +3,5 @@ matplotlib
 numpy
 pandas
 scipy
-sklearn
+scikit-learn
 torch
