+++ flint-data-preprocessing/README.md
@@ -1,7 +1,7 @@
 ## Preprocessing
 
 We preprocess the Flint dataset in the following way:
-(we have also uploaded the processed outputs [procd_velocities.npy](procd_velocities.npy) and [procd_spikes.npy](procd_spikes.npy) directly to this GitHub folder, so you can skip steps 1 and 2 if desired):
+(we have also uploaded the processed outputs [procd_velocities.npy](../procd_velocities.npy) and [procd_spikes.npy](../procd_spikes.npy) directly to the main folder, so you can running the preprocessing steps entirely if you want):
 
 1. Download data [here](https://portal.nersc.gov/project/crcns/download/dream/data_sets/Flint_2012). Place the five .mat files in this directory.
 2. Run [flint_preprocess_data.ipynb](flint_preprocess_data.ipynb) to:
