+++ README.md
@@ -1,6 +1,8 @@
 # Discriminative Kalman-Filter Experiment 4.5 in Python + PyTorch
 Python + PyTorch Reproduction of Experiment 4.5 of the Discriminative Kalman Filter paper from 2020.
 
+NOTE: this code has not been run on a Mac architecture. I ran everything on a PC, so I recommend using `x86_64` architectures.
+
 See [requirements.txt](requirements.txt) for package dependencies.
 
 First, run [flint_preprocess_data.ipynb](flint-data-preprocessing/flint_preprocess_data.ipynb) using the instructions in [its README](flint-data-preprocessing/README.md).
