### 2024-03-05 #74 @Josuelmet
@ozancaglayan I updated the repo a bit, here is my progress:
- [x] Replaced sklearn with scikit-learn in requirements.txt
- [x] Updated preprocessing README to say that running preprocessing is optional and that the processed files should exist in the flint-data-preprocessing folder.
- [x] Switched torch and numpy import order in Paper_Script_45.ipynb
- [x] Updated the main README to suggest using `x86_64` architectures
- [x] Run Paper_Script_45.ipynb in a Colab runtime to make sure code works in a Linux environment.
- [ ] Clear / trim up cell outputs from notebooks.

I will let you know when I have cleared the Linux issues.
Thanks!

Edit: I have cleared the Linux issues in the new code as of 3/6/2024.
