+++ flint-data-preprocessing/README.md
@@ -8,4 +8,4 @@ We preprocess the Flint dataset in the following way:
 * move the data from the separate files and nested structs into a single location,
 * and to coarsen data and hit the neural data with PCA and z-scoring.
 * (After the preprocessing, many sets were only ~6000 pairs long; so each run used 5000 training points and 1000 test points.)
-3. Move [procd_velocities.npy](procd_velocities.npy) and [procd_spikes.npy](procd_spikes.npy) to the main directory.
+* (After preprocessing, [procd_velocities.npy](procd_velocities.npy) and [procd_spikes.npy](procd_spikes.npy) should exist in this directory.)
