### 2024-03-06 80ad2a7
Update nw_est.py
added error context disabler to f_est = xP / sP
