### 2024-03-06 5cb92cf
Undid previous change to nw_est.py
