### 2024-03-06 f209803
Update optimize_sx.py to handle RuntimeWarnings
