### 2024-03-06 3c1dd9b
Update optimize_sx.py to have bracket=(0.1, 1.0) instead of (0.0, 1.0)
