### 2024-03-06 #74 @Josuelmet
@ozancaglayan I have figured out the issue behind the instability, and there are two ways to fix it:
- either use scipy==1.10.1 , which is what I realized I used on my system; or,
- use the newest version of the code/repo (with any version of scipy, including the latest), in which I have tweaked the initialization of an important kernel parameter to avoid instability / nans.

I recommend the latter; let me know when you are able to give running the code another chance.
Thanks!

P.S. I ran the code for one seed (42) in a new Linux environment (Colab); some of the numbers for Table 2 jittered a bit, but overall the trends were the same. The new numbers matched Table 3 very closely, with some jitter in the DKF-NN row.
