### 2024-07-17 #74 @ozancaglayan
Hi,

- Went back to the codebase and I can confirm that the notebook is now running in Linux x86_64 without any issues. I tried with the recent version of `scipy` without pinning it to `1.10.1` and it seems to work fine as well ✅
- All my previous points were addressed by the author ✅

### Remarks about the results in **Table 2**
- Kalman, GP, LSTM and EKF results are exactly as they are given in the Table
- DKF-NW, DKF-NN and UKF have some variations across the trials. What's interesting is that the variations are not happening at each trial but only in some of them, which is a bit weird. Any idea on what may be the reason here? These variations slightly change the trial averages as well. See the below figure where I wrote in red the numbers that I'm getting from your notebook:
(I would have attributed this to some torch/NN related non-determinism but given that LSTM results are exactly the same as the table, I think there's sth else going on here. To verify, I ran the notebook twice, the table differences are consistent, e.g. they don't change across runs)

### Remarks about the results in **Table 3**
- DKF-NN and UKF have some variations again, non-negligible differences for DKF-NN. (Now the variations for DKF-NN happened at each trial. Since these are two different evaluation metrics, I would have expected the diffs also happening in the first table but didn't check whether the scores change slightly all the time or not.)

![image](https://github.com/user-attachments/assets/7285e31d-5008-4a90-b562-dc07db0894e8)

Some minor/further improvements:
- You can set a global precision for rendering the tables: `pd.set_option('display.precision', 3)`
- Could you rename `table1` and `table2` DataFrame's in the notebook to `table2` and `table3` so that they match the table numbers in your paper?
- Could you fix the below issues of setting a string to a float column, which would cause errors in the future so that your pipeline is ensured to be correctly executed by future versions of `pandas`? I think you can simply remove this and multiply the 0-1 improvement scores by 100 and with the `display.precision` fix, they'll show up as `-18.7` for example which should be fine
```python
[/tmp/ipykernel_387075/859728690.py:10](http://10.98.58.43:8001/tmp/ipykernel_387075/859728690.py#line=9): FutureWarning: Setting an item of incompatible dtype is deprecated and will raise an error in a future version of pandas. Value '-19%' has dtype incompatible with float64, please explicitly cast to a compatible dtype first.
  table1.loc[idx] = table1.loc[idx].map(lambda n: '{:.0%}'.format(n))
```

### Conclusion
@Josuelmet I'm curious about your take on the result differences. Thanks!
