### 2024-08-07 55ea9ba
Update DKF_NN.py
add momentum parameter to train_nn
