### 2024-08-07 #74 @Josuelmet
@ozancaglayan I took a deep look at the differences between two runtimes (one on my PC, and one on Colab), and even when using the same version of PyTorch, I was unable to make the two sets of results fully agree, particularly for the simple NN-based architectures: DKF-NN, EKF/UKF. My best guess is that this difference stems from somewhere within PyTorch. The LSTM seems to be more stable since it runs for much fewer epochs than the DKF-NN, EKF, and UKF. Between my two runtimes, I looked at the NN losses, and for some reason they would agree fully up until 330 epochs, after which they diverged. As a result, I tried modifying the training parameters to use fewer epochs, but in the end I was unable to fully remove all discrepancies between my two runtimes. Let me know what you think. Thanks!
