### 2024-12-05 #74 @ozancaglayan
I had some minor remarks in my last comment. Besides the inconsistencies in a small part of the results that we discussed, I'm fine with this, sorry for the delays!
