### 2025-02-17 #74 @rougier
We're very very late on this submission and based on the thorough review by @ozancaglayan and subsequent modifications by @Josuelmet I propose the article to be accepted. @Josuelmet (congrats !). Can you tell me where are your article sources? (because I'll need to produce the final PDF)
