### 2025-02-25 #74 @Josuelmet
Thanks @rougier @ozancaglayan !
I made everything on Overleaf, attached is the .zip source from the project.
Let me know if there's anything else you need from me, or if you want to access the Overleaf project directly.

[\[Re\] The Discriminative Kalman Filter for Bayesian Filtering with Nonlinear and Nongaussian Observation Models.zip](https://github.com/user-attachments/files/18969421/Re.The.Discriminative.Kalman.Filter.for.Bayesian.Filtering.with.Nonlinear.and.Nongaussian.Observation.Models.zip)
