### 2025-03-24 #74 @rougier
Thanks. The sandbox version (not the final one) is published here:
https://sandbox.zenodo.org/record/187904

Can you check everything is ok (links, name, ref, etc) ?
