### 2025-04-08 #74 @rougier
It's online at https://zenodo.org/records/15172014. It will soon appear on the ReScience website. Congratulations!
